RGFdata: Extração intuitiva e fácil dos dados do RGF

Realiza a extração dos dados do Relatório de Gestão Fiscal (RGF) de maneira mais intuitiva e fácil utilizando a API do SICONFI.

Usage

RGFdata(
  cod.ibge = NULL,
  year = NULL,
  power = NULL,
  period = NULL,
  annex = NULL,
  simplified = FALSE
)

Argumentos

cod.ibge

Código IBGE do Ente. se simplified = TRUE, então selecione municípios. Para obter todas as UF's utilize "all_states". Para obter todos os municípios utilize "all_muni".

year

Exercício do relatório

power

Código do poder. Valores disponíveis: E = Executivo, L = Legislativo, J = Judiciário, M = Ministério Público, D = Defensoria Pública

period

Quadrimestre ou semestre de referência do relatório dentro de um exercício. A periodicidade semestral é automaticamente selecionada se simplified = TRUE. A periodicidade padrão do relatório é quadrimestral. Valores disponíveis: 1, 2 ou 3.

annex

Anexos dos demonstrativos do RGF. Valores disponíveis: 1, 2, 3, 4, 5 ou 6.

simplified

Tipo do Demonstrativo. RGF Simplificado aplica-se apenas aos municípios com menos de 50 mil habitantes que optaram pela publicação semestral dos relatórios. Se TRUE a periodicidade semestral será automaticamente selecionada.

Exemplos

# install.packages("devtools")
devtools::install_github("Natanaelsl/RREORGFdataR")
#> Using github PAT from envvar GITHUB_PAT. Use `gitcreds::gitcreds_set()` and unset GITHUB_PAT in .Renviron (or elsewhere) if you want to use the more secure git credential store instead.
#> Downloading GitHub repo Natanaelsl/RREORGFdataR@HEAD
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/Rtmpvljdkl/remotes16e9f7f669c/Natanaelsl-RREORGFdataR-ba273ca/DESCRIPTION’ ... OK
#> * preparing ‘RREORGFdataR’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘RREORGFdataR_0.1.0.9000.tar.gz’
#> 
#> Installing package into ‘/home/runner/work/_temp/Library’
#> (as ‘lib’ is unspecified)

# Carregando o pacote
library(RREORGFdataR)

# Extraindo dados do anexo 1 para o 3º quadrimestre
# do RGF de 2020 até 2023 do Estado de Goiás para todos os poderes.
RGFdata(cod.ibge = 52,
        year = c(2020:2023),
        power = c('E','L','J','D','M'),
        period = 3,
        annex = 1,
        simplified = FALSE)
#> Warning: internal error 1 in R_decompress1 with libdeflate
#> Error in eval(expr, envir, enclos): lazy-load database '/home/runner/work/_temp/Library/RREORGFdataR/R/RREORGFdataR.rdb' is corrupt