RREORGFdataR: Extração intuitiva e fácil dos dados do RREO e RGF utilizando API do SICONFI.

O 'RREORGFdataR' é um pacote R que permite aos usuários acessar facilmente o conjunto de dados do Relatório Resumido da Execução Orçamentária (RREO) e do Relatório de Gestão Fiscal (RGF) utilizando a API do SICONFI. O pacote foca em alta performance e automação de extração para analistas da área fiscal.

Details

O pacote abstrai a complexidade da API do Tesouro Nacional, implementando funções para extração em lote, tratamento de erros com retry automático e estruturação otimizada para análise em R.

See also

Author

Natanael Soares Leite natanael.leite@goias.gov.br Raphael Maciel de Lima