RREORGFdataR: Extração intuitiva e fácil dos dados do RREO e RGF utilizando API do SICONFI.

O 'RREORGFdataR' é um pacote R que permite aos usuários acessar facilmente o conjunto de dados do Relatório Resumido da Execução Orçamentária (RREO) e do Relatório de Gestão Fiscal (RGF) utilizando a API do SICONFI (https://apidatalake.tesouro.gov.br/docs/siconfi/). Saiba mais sobre a utilização do 'RREORGFdataR' em https://natanaelsl.com.br/project/rreorgf_pkg/ ou https://natanaelsl.github.io/RREORGFdataR/.

Detalhes

O Tesouro Nacional disponibilizou a Application Programming Interface (API) de dados abertos para atender à demanda por dados brutos oriundos do Sistema de Informações Contábeis e Fiscais do Setor Público Brasileiro – Siconfi. Por meio dessa ferramenta, é possível ao usuário obter desde pequenas frações até grandes volumes de dados de todas as informações inseridas pelos entes subnacionais no Siconfi.

Veja também

Autor

Natanael Soares Leite (ORCID)

Raphael Maciel de Lima